requirements.txt
psycopg2-binary==2.9.6
pandas==1.5.3
configparser==5.3.0
boto3==1.26.90
python-dotenv==1.0.0